Fixed bug with breaks
Breaks were getting added to the database even
though they are not pomodoros.

I changed this adding some event details when the
end event gets triggered. The end event has the mode
at the time of end so we can use that to determine
if we should tell the server to store a new pomodoro
or not.