Fixed tests and bug and added logger
I added a logger to figure out when date gets misaligned so that
I can fix it.

All the tests were breaking...so I fixed that.

Turns out default arguments in Python functions are only computed
at the start of the program. If your default argument is mutable it
will not be updated. This is what caused the issue with the website
crashing after midnight and is now fixed.